Support CLCACHE_BASEDIR in nodirect mode
Do a simple case-insensitive find-and-replace to transform absolute
paths into relative paths within the preprocessor output that is used
to compute the hash. This makes CLCACHE_NODIRECT mode usable in the
presence of the __FILE__ macro.

Fixes a bug in _normalizedCommandLine that caused the source filename
to be included in the hash computation.